extensions/claude-chat/pyproject.toml
[project]
name = "claude-chat"
version = "0.0.0"
description = "A basic chat interface powered by the Claude Agent SDK"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "shiny>=1.4.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]
